create ShardsAffectedByTeamFailure.h
